+++ src/RaceProofServiceProvider.php
@@ -10,11 +10,15 @@ use RaceProof\Laravel\Console\CleanCommand;
 use RaceProof\Laravel\Console\DoctorCommand;
 use RaceProof\Laravel\Console\WorkerCommand;
 use RaceProof\Laravel\Contracts\CoordinatorStore;
+use RaceProof\Laravel\Contracts\RaceClock;
 use RaceProof\Laravel\Contracts\RequestExecutor;
+use RaceProof\Laravel\Contracts\WorkerProcessFactory;
 use RaceProof\Laravel\Coordination\FileCoordinatorStore;
 use RaceProof\Laravel\Execution\KernelRequestExecutor;
 use RaceProof\Laravel\Execution\RaceContext;
+use RaceProof\Laravel\Execution\SymfonyWorkerProcessFactory;
 use RaceProof\Laravel\Support\ConfigValue;
+use RaceProof\Laravel\Support\SystemRaceClock;
 
 final class RaceProofServiceProvider extends ServiceProvider
 {
@@ -29,6 +33,8 @@ final class RaceProofServiceProvider extends ServiceProvider
 
         $this->app->singleton(RaceContext::class);
         $this->app->singleton(RacePoint::class);
+        $this->app->singleton(RaceClock::class, SystemRaceClock::class);
+        $this->app->singleton(WorkerProcessFactory::class, SymfonyWorkerProcessFactory::class);
         $this->app->bind(RequestExecutor::class, KernelRequestExecutor::class);
         $this->app->bind(RaceBuilder::class);
     }